primary key error

+++ Curso_Java_a_.net/DataAccess/Entities/Subjects.cs
@@ -2,7 +2,7 @@ namespace Curso_Java_a_.net.DataAccess.Entities
 {
     public class Subjects
     {
-        public int SubjectId { get; set; }
+        public long SubjectId { get; set; }
         public string Name { get; set; }
         public string Schedule { get; set; }
     }
