+++ Curso_Java_a_.net/DataAccess/Entities/UsersSubjects.cs
@@ -2,9 +2,9 @@ namespace Curso_Java_a_.net.DataAccess.Entities
 {
     public class UsersSubjects
     {
-        public int SubjectIdUsuario { get; set; }
-        public int SubjectId { get; set; }
-        public int UserId { get; set; }
+        public long SubjectIdUsuario { get; set; }
+        public long SubjectId { get; set; }
+        public long UserId { get; set; }
         public int Grade { get; set; }
     }
 }
